refactor: rename test wrapper
